[Refactor] refactoring for CVRPTW
